[Chore] #21 - ProfileView 위치 이동
